refactor(tools): separate HTTP-related functions from tool.utils
KAG-2957
