+++ kong-3.6.0-0.rockspec
@@ -175,6 +175,7 @@ build = {
     ["kong.tools.time"] = "kong/tools/time.lua",
     ["kong.tools.module"] = "kong/tools/module.lua",
     ["kong.tools.ip"] = "kong/tools/ip.lua",
+    ["kong.tools.http"] = "kong/tools/http.lua",
 
     ["kong.runloop.handler"] = "kong/runloop/handler.lua",
     ["kong.runloop.events"] = "kong/runloop/events.lua",
